Adding Color AND Updating ReadME file
Adding color to the map making it look more like a terrain map. Doing this by utilizing our previously added in version of height for our vertex which was noise_image.get_pixel(x, y).r * 64, otherwise known as essentially the height and coding it into some usable variables such as  'heightAdjustedColor' and 'z'. Using heightAdjustedColor the overall color of the entire landscape is adjusted as per each individual height peak on the map as to individualize the triangles by color to form a general looking landscape from above, complete with all the basics. Not much of a change from the regular code besides customized color and a variable touch up. Added this into the README file afterwards to further explain each addition of color and more.